+++ extra-bonus-notifier/extra_bonus_notifier.go
@@ -198,7 +198,7 @@ func (ebn *extraBonusNotifier) notifyingExtraBonusAvailability(ctx context.Conte
 
 		for _, usr := range userResults {
 			var extraBonusIndex uint16
-			if isAvailable, _ := IsExtraBonusAvailable(now, ebn.extraBonusStartDate, usr.ExtraBonusStartedAt, ebn.extraBonusIndicesDistribution, usr.ID, usr.UTCOffset, &extraBonusIndex, &usr.ExtraBonusDaysClaimNotAvailable, &usr.ExtraBonusLastClaimAvailableAt); isAvailable {
+			if isAvailable, _ := IsExtraBonusAvailable(now, ebn.extraBonusStartDate, usr.ExtraBonusStartedAt, ebn.extraBonusIndicesDistribution, usr.ID, int16(usr.UTCOffset), &extraBonusIndex, &usr.ExtraBonusDaysClaimNotAvailable, &usr.ExtraBonusLastClaimAvailableAt); isAvailable {
 				eba := &ExtraBonusAvailable{UserID: usr.UserID, ExtraBonusIndex: extraBonusIndex}
 				updatedUsers = append(updatedUsers, &usr.UpdatedUser)
 				msgs = append(msgs, ExtraBonusAvailableMessage(ctx, eba))
